Move document utils to seperate files
Again, due to later changes, this was causing a dependency cycle so I
think it’s wise to move these functions to their own file. I have
created an assessments folder and I suggest we move other assessment
utils out to this folder piece by piece
