+++ server/utils/applicationUtils.test.ts
@@ -5,16 +5,8 @@ import Apply from '../form-pages/apply'
 import { DateFormats } from './dateUtils'
 import { tierBadge } from './personUtils'
 
-import {
-  getResponses,
-  getPage,
-  getArrivalDate,
-  dashboardTableRows,
-  documentsFromApplication,
-  overwriteApplicationDocuments,
-} from './applicationUtils'
+import { getResponses, getPage, getArrivalDate, dashboardTableRows } from './applicationUtils'
 import { SessionDataError, UnknownPageError } from './errors'
-import documentFactory from '../testutils/factories/document'
 
 const FirstPage = jest.fn()
 const SecondPage = jest.fn()
@@ -158,44 +150,6 @@ describe('applicationUtils', () => {
           },
         ],
       ])
-    })
-  })
-  describe('documentsFromApplication', () => {
-    it('returns the selected documents if they exist', () => {
-      const application = applicationFactory.build()
-      const documents = documentFactory.buildList(2)
-
-      application.data['attach-required-documents'] = {
-        'attach-documents': {
-          selectedDocuments: documents,
-        },
-      }
-
-      expect(documentsFromApplication(application)).toEqual(documents)
-    })
-
-    it('returns an empty array if the application doesnt have selected documents', () => {
-      const application = applicationFactory.build()
-
-      expect(documentsFromApplication(application)).toEqual([])
-    })
-  })
-
-  describe('overwriteApplicationDocuments', () => {
-    it('overwrites the current documents on the application with the ones supplied and returns the application', () => {
-      const application = applicationFactory.build()
-      const originalDocuments = documentFactory.buildList(1)
-      const newDocuments = documentFactory.buildList(2)
-
-      application.data['attach-required-documents'] = {
-        'attach-documents': {
-          selectedDocuments: originalDocuments,
-        },
-      }
-
-      const applicationWithNewDocuments = overwriteApplicationDocuments(application, newDocuments)
-
-      expect(documentsFromApplication(applicationWithNewDocuments)).toEqual(newDocuments)
     })
   })
 })
